virtue-ai-service/requirements.txt
flask==3.0.0
google-cloud-aiplatform==1.32.0
google-cloud-vertexai==1.9.0
gunicorn==21.2.0